requirements.txt
tensorflow-gpu>=1.9.0
opencv-python
scipy
scikit-image
h5py
git+git://github.com/alexlee-gk/lpips-tensorflow.git#egg=lpips-tf
